+++ src/dping/options.cpp
@@ -29,16 +29,16 @@ Usage: )" FILE_INTERNAL_NAME R"( --help
     --version
     --list-country [-v] [--debug]
     --list-asn code [-v] [--debug]
-    --country code [-n probes] [-w timeout] [-wa timeout] [--no-delays] [-v]
-                   [--debug] {target_name}
-    --asn id [-n probes] [-w timeout] [-wa timeout] [--no-delays] [-v]
-             [--debug] {target_name}
+    --country code [-n probes] [-w timeout] [-t] [-v] {target_name}
+    --asn id [-n probes] [-w timeout] [-t] [-v] {target_name}
 
 Options:
     {target_name}  Destination host IP or domain name.
 
     --help          Display this help.
     --version       Display detailed program version, copyright notices.
+    --list-country  List available countries.
+    --list-asn code List ASNs for specified 2 letter country code.
     --country code  Specify source addresses 2 letter country code
                     (ISO 3166-1 alpha-2).)"
 #ifdef DO_BY_COUNTRY_BY_DEFAULT
@@ -47,20 +47,31 @@ R"(
                     available is a default setting.)"
 #endif
 R"(
-    --asn id        Use source addresses from specified ASN
+    --asn id        Use source addresses from specified ASN.
                     (autonomous system number) network.
     -n probes       Probes limit: number of hosts to make requests from.
+                    This option has an alias: --probes probes
     -w timeout      Timeout in milliseconds to wait for single ping.
+    -t              Ping the specified host until stopped.
+                    To see statistics and continue - type Control-Break;
+                    To stop - type Control-C.
+    -v              Verbose output.
+
+Advanced options:
     -wa timeout     Timeout in milliseconds to wait for all probes.
-    --list-country  List available countries.
-    --list-asn code List ASNs for specified 2 letter country code.
+    -a              Resolve addresses to hostnames.
+    -np count       Number of pings per probe to send (default: 1).
+    -i TTL          Time To Live.
+    -l size         Send buffer size.
+    -f              Set Don't Fragment flag in packet (IPv4-only).
+    -4              Force using IPv4.
+    -6              Force using IPv6.
+    --api-key key   Set web API key.
+    --api-url url   Set web API URL.
     --no-delays     Disable delays during printing of results to console.
-    --api-key key   Set web API key
-    -v              Verbose output
-    --debug         Additional debug output
+    --debug         Additional debug output.
+
 )"
-+ (options.bDebug ? GetDebugArgumentsHelp() : "")
-+ "\n"
 + GetReturnCodeInfo()
 + R"(
 Examples: