+++ src/dtraceroute/options.h
@@ -37,10 +37,16 @@ public:
 	bool		bVerbose		= false;
 	bool		bDebug			= false;
 	bool		bNoDelays		= false;
+
+	bool		bResolveIp2Name	= true;		// this is true in Windows by default
+	bool		bUseIpv4Only	= false;
+	bool		bUseIpv6Only	= false;
+	bool		bDontFragment	= false;
+
 	uint32_t	nTimeoutPingMs	= 1000;
 	uint32_t	nTimeoutTotalMs	= 30 * 1000;
-	const uint32_t nWaitBetweenPingsMs = 100;
-	const uint32_t nTriesPerHop = 3;
+	uint32_t	nWaitBetweenPingsMs = 100;
+	uint32_t	nTriesPerHop	= 3;
 	uint32_t	nProbesLimit	= 10;
 	uint32_t	nResultsLimit	= UINT32_MAX;		// for list ASNs and list countries modes; UINT32_MAX - display ALL items from requested list
 	uint16_t	nPacketSize		= 32;